# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MLOVI - Multidisciplinar: BACKEND API V1",
        "description": "Description of your API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Main (production) server"
        }
    ],
    "paths": {
...
        "/lotes": {
            "get": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Listar todos as lotes",
                "operationId": "getLoteList",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    }
                }
            },
            "post": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Cadastrar uma nova lote",
                "description": "Cria um novo registro de lote com os dados fornecidos.",
                "operationId": "storeLote",
                "requestBody": {
                    "description": "Dados da lote",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_convenio",
                                    "numero",
                                    "tipo",
                                    "data",
                                    "status",
                                    "valor_total"
                                ],
                                "properties": {
                                    "id_convenio": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "1"
                                    },
                                    "tipo": {
                                        "type": "string",
                                        "example": "Consulta"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-30"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "0"
                                    },
                                    "valor_total": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lote criado com sucesso"
                    },
                    "400": {
                        "description": "Erro na requisição"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/lotes/{id}": {
            "get": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Buscar lote por ID",
                "operationId": "getLoteById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lote",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Atualizar lote",
                "operationId": "updateLote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lote",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Dados do lote a serem alterados. Verifique os campos no Schema",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Lote Ilha do leite"
                                    },
                                    "endereco": {
                                        "type": "string",
                                        "example": "Av. Frei Matias Teves, 280 - Ilha do Leite, Recife - PE, 50070-465"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Excluir lote",
                "operationId": "deleteLote",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do lote",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Lote excluído com sucesso"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    }
                }
            }
        },
        "/lotes/{idConvenio}/{numeroLote}/xml": {
            "get": {
                "tags": [
                    "Lotes"
                ],
                "summary": "Gerar XML do lote de um convênio",
                "operationId": "generateXML",
                "parameters": [
                    {
                        "name": "idConvenio",
                        "in": "path",
                        "description": "ID do Convênio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "numeroLote",
                        "in": "path",
                        "description": "Número do lote atrelado ao convênio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    },
                    "401": {
                        "description": "Não Autenticado"
                    },
                    "403": {
                        "description": "Proibido"
                    },
                    "404": {
                        "description": "Lote não encontrado"
                    }
                }
            }
        },
...
    },
...
}